webserver

+++ esb/src/main/java/com/team7/esb/dto/BankProxyResponseDTO.java
@@ -1,10 +1,8 @@
 package com.team7.esb.dto;
 
-import java.io.Serializable;
-
 public class BankProxyResponseDTO {
     public String uid;
-    public int interest;
+    public float interest;
     public int initialPayment;
     public String bankName;
 }
